+++ components/scaletor/Layout/index.js
@@ -14,6 +14,7 @@ import SelectorDouble from '../SelectorDouble';
 import NotesHistory from '../NotesHistory';
 
 const Layout = () => {
+  const [isMobile, setIsMobile] = useState(false);
   const router = useRouter();
   const { query } = router;
 
@@ -178,6 +179,22 @@ const Layout = () => {
     });
   }, [scale]);
 
+  // Resize
+  useEffect(() => {
+    if (typeof window !== 'undefined') {
+      const handleResize = () => {
+        setIsMobile(window.innerWidth <= 1024);
+      };
+
+      handleResize(); // Set the initial value
+
+      window.addEventListener('resize', handleResize);
+      return () => {
+        window.removeEventListener('resize', handleResize);
+      };
+    }
+  }, []);
+
   /**
    * Return
    */
@@ -195,6 +212,7 @@ const Layout = () => {
             options={tunningOptions}
             onChange={handleTunningChange}
             value={initialTunning}
+            isMobile={isMobile}
           />
         </div>
         <div className="layout-scaletor__column-3">
@@ -210,6 +228,7 @@ const Layout = () => {
             checkedScaleSwitch={scaleSwitch}
             valueNote={initialScaleNote}
             valueScale={initialScaleName}
+            isMobile={isMobile}
           />
         </div>
         <div className="layout-scaletor__column-3">
@@ -232,15 +251,10 @@ const Layout = () => {
             ]}
             onChange={handleTemplateChange}
             value={initialTemplate}
+            isSearchable={false}
+            isMobile={isMobile}
           />
         </div>
-      </div>
-      <div className="layout-scaletor__bottom">
-        <div className="layout-scaletor__column-3">
-          <NotesHistory />
-        </div>
-        <div className="layout-scaletor__column-3" />
-        <div className="layout-scaletor__column-3" />
       </div>
       <div className="layout-scaletor__center">
         <Matrix
@@ -253,6 +267,9 @@ const Layout = () => {
           withoutScale={!scaleSwitch}
         />
         <Template id={template} />
+      </div>
+      <div className="layout-scaletor__bottom">
+        <NotesHistory />
       </div>
     </section>
   );