+++ pages/home/index.js
@@ -4,7 +4,6 @@ import pkg from '../../package.json';
 
 import Tools from '../../components/home/Tools';
 import Footer from '../../components/commons/Footer';
-import Unsupported from '../../components/commons/Unsupported';
 
 const footer = {
   project: {
@@ -29,7 +28,6 @@ const HomePage = () => (
         project={footer.project}
         back={footer.back}
       />
-      <Unsupported />
     </main>
   </>
 );