+++ structures/quintor/CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unchanged
+
+## [1.1.0] - 2024-08-28
+### Changes
+- Add responsive to all pages and components.
+
 ## [1.0.1] - 2022-11-05
 ### Changes
-- Update Fifth Value Selection [#10](https://github.com/fdrandolfi/composition-tools-frontend/pull/10)
+- Update Fifth Value Selection.
 
 ## [1.0.0] - 2022-11-01
 ### Changes