+++ structures/scaletor/CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unchanged
 
+## [1.13.0] – 2024-08-28
+### Changes
+- Add responsive to all pages and components.
+
 ## [1.12.0] – 2024-08-25
 ### Changes
 - Add new templates for PRS and Dean guitars, update responsive styles and add general changes.